+++ tensornetwork/block_sparse/__init__.py
@@ -2,9 +2,14 @@ from tensornetwork.block_sparse import index
 from tensornetwork.block_sparse import charge
 from tensornetwork.block_sparse import blocksparsetensor
 from tensornetwork.block_sparse import linalg
-#pylint: disable=line-too-long
-from tensornetwork.block_sparse.blocksparsetensor import BlockSparseTensor, ChargeArray, tensordot, outerproduct
-#pylint: disable=line-too-long
-from tensornetwork.block_sparse.linalg import svd, qr, diag, sqrt, trace, inv, pinv, eye, zeros, ones, randn, eigh, eig, conj, reshape, transpose, random, norm
+from tensornetwork.block_sparse.blocksparsetensor import (BlockSparseTensor,
+                                                          ChargeArray,
+                                                          tensordot,
+                                                          outerproduct)
+from tensornetwork.block_sparse.linalg import (svd, qr, diag, sqrt, trace, inv,
+                                               pinv, eye, zeros, ones, randn,
+                                               eigh, eig, conj, reshape,
+                                               transpose, random, norm)
 from tensornetwork.block_sparse.index import Index
-from tensornetwork.block_sparse.charge import U1Charge, BaseCharge, Z2Charge, ZNCharge
+from tensornetwork.block_sparse.charge import (U1Charge, BaseCharge, Z2Charge,
+                                               ZNCharge)
