+++ tensornetwork/block_sparse/charge.py
@@ -703,42 +703,6 @@ def fuse_charges(charges: List[BaseCharge], flows: List[bool]) -> BaseCharge:
   return fused_charges
 
 
-def fuse_degeneracies(degen1: Union[List, np.ndarray],
-                      degen2: Union[List, np.ndarray]) -> np.ndarray:
-  """
-  Fuse degeneracies `degen1` and `degen2` of two leg-charges
-  by simple kronecker product. `degen1` and `degen2` typically belong to two
-  consecutive legs of `BlockSparseTensor`.
-  Given `degen1 = [1, 2, 3]` and `degen2 = [10, 100]`, this returns
-  `[10, 100, 20, 200, 30, 300]`.
-  When using row-major ordering of indices in `BlockSparseTensor`,
-  the position of `degen1` should be "to the left" of the position of `degen2`.
-  Args:
-    degen1: Iterable of integers
-    degen2: Iterable of integers
-  Returns:
-    np.ndarray: The result of fusing `dege1` with `degen2`.
-  """
-  return np.reshape(
-      np.array(degen1)[:, None] * np.array(degen2)[None, :],
-      len(degen1) * len(degen2))
-
-
-def fuse_ndarrays(arrays: List[Union[List, np.ndarray]]) -> np.ndarray:
-  """
-  Fuse all `arrays` by simple kronecker addition.
-  Arrays are fused from "right to left", 
-  Args:
-    arrays: A list of arrays to be fused.
-  Returns:
-    np.ndarray: The result of fusing `arrays`.
-  """
-  if len(arrays) == 1:
-    return np.array(arrays[0])
-  fused_arrays = np.asarray(arrays[0])
-  for n in range(1, len(arrays)):
-    fused_arrays = np.ravel(np.add.outer(fused_arrays, arrays[n]))
-  return fused_arrays
 
 
 def charge_equal(c1: BaseCharge, c2: BaseCharge) -> bool:
